tests: use standard CONFIG+=testcase for `make check'
Remove the custom check targets from jsondb tests and replace them with
the standard check target from CONFIG+=testcase, which is more powerful.
This allows the tests to be run like the tests for all other Qt modules
(e.g. with backtraces on crash and with accurate log capturing).

For the benchmarks, `check' is removed entirely, since running
benchmarks on a desktop generally isn't useful.  This is consistent with
the behavior of benchmarks in qtbase.

If necessary, the behavior of the old `check' commands can be reproduced
with something like:

  make -j1 check \
    TESTRUNNER="rm -f *.db* && LD_LIBRARY_PATH=$PWD/lib QT_QPA_PLATFORM=xcb " \
    TESTARGS="-xunitxml -o testlog.xml -silent"

Change-Id: Icf8bf2cf0613e3e8b21aa4f206ce3d76559908f7
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jamey Hicks <jamey.hicks@nokia.com>
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>