Implemented Map with new join option.
With the new join option, join is symmetric in the source types. A
change in any source type will cause the corresponding function to be
invoked and to update the target type appropriately.

Updated documentation and examples.

The old Map with lookup behavior is deprecated and will be removed.

Change-Id: Ic722ed42834eec4253feb26130858a32e682cdf0
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Kevin Simons <kevin.simons@nokia.com>
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>