Enable apps with "setOwner" capability to set _owner field.
Applications such as backup/restore and the installer need to be able
to create objects with ownership other than their own. This change
leaves the "_owner" property unchanged if it is set and if the
connected application has the "setOwner" access on the object being
updated.

Change-Id: I2d4a11a4eed9d61cef165b4c7945039fd04766a6
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>