Fixed problem reindexing properties.
ObjectTable::reindexObjects() iterates over they keys in the object
table. It was skipping those whose key was a state key but it was
mistakenly indexing objects in the history, whose key is a state key
prefixed with an object key.

Object keys are 16 bytes, state keys are 5 bytes, and history entry
keys are 21 bytes.

It might be better to use a byte or two at the beginning of the
key to distinguish them, rather than the length.

Change-Id: Idc81f5ba69183ee793a8d5d17d54577e409f313b
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>