Polished JsonDbClient behavior
Made the connection to the qtjsondb server completely transparent for users of
the api - i.e. if there is no jsondb daemon running at this time, all requests
(create/update/remove/query commands) will be queued in JsonDbClient and sent
to the server when that becomes available.

At the same time, JsonDbClient tries to connect to the server every 5 seconds
if the connection was dropped for any reason (e.g. imaging qtjsondb crashes and
is restarting itself - with this patch that will be transparent to users and
requests will be reset after qtjsondb is back online)

Change-Id: If3683c5c1fbac672c481ac3aeecf8e7a251c9211
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>