Added toggle for key sizes
There is a theoretical corner case which requires a change in
the split algorithm, so until that's figured out, we should
stick with 255 byte keys. Will try and produce it and then
decide what to do:

The situation is that if you had 3 keys in a page and your
have to insert k5:

|k1|k2|k3|k4|

If the k5 needs be appended and the split happens between
k2 and k3, then k1 and k2 go on a new left page while k3 and
k4 go on a new right page. Seems like k5 could be too big
for the right page since we guarantee 2 keys per page, so we'd
need a triple split algorithm.

Change-Id: If30c59cc1b86ac2330501df83e75f27ac3f6cb07
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jamey Hicks <jamey.hicks@nokia.com>
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>