Implemented ephemeral partition
Objects in the ephemeral partition are not persisted to disk but stored in memory instead.
Less operations are allowed on those objects, e.g. one cannot query them with sort order,
version checking is not enforced on them, there are no indexes and views in that partition.

Notification objects are stored in the ephemeral partition.

Change-Id: Id3efd51ddd4dbc3542e6dc7facaea2ed6639bc97
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>
Sanity-Review: Denis Dzyubenko <denis.dzyubenko@nokia.com>