marker bug fix + test
When a file is opened and written to and then killed,
the next open does not work because it expects a
marker is set since the file is bigger than just the
header.

The static functions in btree_p.h were moved in to
btree.cpp as well since they are not used yet in
other places internally.

Change-Id: Ifba6fa0305d72c69d0cf03659af47fc74d042791
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>