Use QJSValue::call() overload that doesn't require "this" argument
The old call() function requires you to explicitly pass a "this"
argument if you want to pass any arguments to the function. The
new call() function only takes an argument list, and uses the
Global Object as the "this" object by default.

The function QJSValue::callWithInstance() can be used to call a
function with a custom "this" object.

Change-Id: I256e3467b81a1f92d4079df835b311b856d5672f
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>