Ensure that only one auto reconnect timer is started.
The auto reconnect can cause multiple single shot timeout timers to be
started.  This eventually leads to an exponential increase in the number
of active timers.  Once the number of active timers reaches tens of
thousands all event processing grinds to a halt causing the application
to freeze.  This behaviour is easily seen when the jsondb daemon is not
running and is never started.

Change-Id: Ia51c93436377570119515650bf2c0ebd4b93cf05
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>