Test change in notification query match as create or remove on update().
Fixes bug #198.

Treat change in query match as create or remove as appropriate even if
object was only updated. For the purposes of notification, if the
query changes from not matching to matching consider it to be a
"create". If it changes from not matching to matching consider it to
be a "remove".

Change-Id: I67881bfe65a0ca9a9ccf98a0d9a5d3a520412d33
Reviewed-by: Kevin Simons <kevin.simons@nokia.com>