Added a control for transaction size to jsondb daemon
This enables a createList, updateList, or removeList operation to be
split into multiple transactions in order to reduce memory
requirements.

All of the modified btree pages are held in DRAM until the transaction
is committed. Because _uuid is random, there is no locality with the
btree. In my measurements, I found that the number of pages involved
in the transaction varied between 0.5 pages per object to 1 page per
object, where there were fewer pages per object with larger
transactions. That does reduce the number of pages written but if DRAM
is limited it may still be better to use perform more writes rather
than to use more DRAM.

Change-Id: I557cde939802ed0cd4130407746b4a649dbda67b
Reviewed-by: Ali Akhtarzada <ali.akhtarzada@nokia.com>