Further enhancements to write access control support
Binding also supported at the left hand side of a query term.
Binding is evaluated as a literal so it is now possible to
have queries like [?%owner startsWith "com.foo."].

In updates the isAllowed is called for both
the existing object and the to-be-new one. This is to ensure
that user has rights to write both. Especially important
in case of updating the _type.

Change-Id: I33f30ebfeb369f00aa17880ec25fcfafffe82df3
Reviewed-by: Jamey Hicks <jamey.hicks@nokia.com>