Add hbtree to daemon
Added an implementation for a new btree backend and a compile time
toggle to switch between the current append only btree and the new
one. Define JSONDB_USE_HBTREE to compile the daemon with the new
btree.

JsonDbManagedBtree was replaced with JsonDbBtree so that new btree
implementations can be dropped in. Also added the use of the context
field in the btree's comparison funciton so that we can use standard
compare functions that use QByteArrays instead.

All references to QBtreeCursor, QBtreeTxn and QBtree were changed to
JsonDbBtree::Cursor, JsonDbBtree::Transaction and JsonDbBtree
respectively. Any new Btree class just has to provide a typedef
for it's cursor type and transaction type. The appropriate functions
must be fixed up in JsonDbBtree as well to account for the different
api of any new btree.

Changed QBtree so that it uses the new compare function and changed
QBtreeCursor::prev to QBtreeCursor::previous

Change-Id: I3e5603a2ba58ac0e81b8ee111e4f6dd1758cdb17
Reviewed-by: Ali Akhtarzada <ali.akhtarzada@nokia.com>
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>