Refactoring: remove JsonDb class
This is the first major step in separating out the partition
functionality into its own library. It completely removes the JsonDb
class and splits the functionality between DBServer and JsonDbPartition.

The goal of this patch is to minimize the changes to existing code. To
that end, a great deal of code was copied-and-pasted from jsondb.cpp.
Subsequent patches will further the refactoring, removing the redundant
code created by this patch.

Change-Id: I33347b802d53c00375829f1a26277bb142a272da
Reviewed-by: Jamey Hicks <jamey.hicks@nokia.com>