Watching a view type triggers a view update.
Registering QJsonDbWatcher on a view type triggers an update of the view.

Rebased onto 'Fixed qjsondbwatcher autotest'.

Fixes Bug #6707

If a watcher is registered on a view type, trigger an update of the
view if the view has not been updated since initialStateNumber, where
initialStateNumber of zero is the current state of the partition.

Change-Id: I55f39f67f4a08f1079ad674d1d3e0119a8ea6ac9
Reviewed-by: Kevin Simons <kevin.simons@nokia.com>