Only index specified objectTypes.
If objectTypes are specified in an Index definition, then only index
objects whose type is contained in objectTypes in this index.

Removed unused index created by JsonDbReduceDefinition because it
caused conflicting indexes to be created.

Fixed non-conforming schema "extends" syntax for the Map/Reduce used
in an affected auto test.

Fixes Bug #6216.

Change-Id: Ic3bda2e7541093b0b5fec42545a3ef9465a3745f
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>