Fix "notExists" query on indexed properties.
If the query engine chooses to use an index on property "x" for a
query with a "notExists" constraint on that property, the result set
will always be empty. This change notes which indexes cannot be used
due to "notExists" and chooses a different one to execute the query.

Fixes Bug #7182.

Change-Id: I15a0a998a0e1dce345d7056167700c0acbff1255
Reviewed-by: Kevin Simons <kevin.simons@nokia.com>