Default deterministic _uuid for Map views.
If the map function does not assign _uuid, then create _uuid for the
emitted objects as follows.

If targetKeyName is not provided:
    sourceUuids.sort()
    identifier = targetType + ":" + sourceUuids.join(":")
If targetKeyName is provided
    sourceUuids.sort()
    identifier = targetType + ":" + sourceUuids.join(":") + ":" + object[targetKeyName].toString()

    object[_uuid] = createUuidFromString(identifier).toString()

If targetKeyName is not provided and _uuid is not assigned by the map
function, then there will be at most once view object per source
object.

Fixes Bug #6807 and Bug #8226

Change-Id: I000755a8a14e5842c1d1b52d9c12a64021ad0790
Reviewed-by: Kevin Simons <kevin.simons@nokia.com>