Includes the values of a named index in query results.
If a named index is used in a query, then it includes the value of the
index in the result. This is useful in the case of collation or
indexes defined by JavaScript propertyFunctions.

For example:
    {
        "_type": "Index",
        "name": "myIndex",
        ...
    }

Given a query:
    [/myIndex]

Each result will have the index value for the object inserted as
property "_indexValue".

In particular, this provides a way for the JsonDb list models to fetch
the index values without having to recompute them.

Change-Id: I6c9b9ea26e1a85b08786c15c0801c8255496cdec
Reviewed-by: Prasanth Ullattil <prasanth.ullattil@nokia.com>