Switch to a partition-loading system based on files
The database searches through a set of paths looking for files named
partitions*.json. These files contain a list of objects specifying
the name and paths of partitions.

The search path is defined as follows:
1) If a path is passed to jsondb, it is the first path searched,
   otherwise the current working directory is first
2) Next is the path specified by the environment variable
   JSONDB_CONFIG_SEARCH_PATH
3) Next is /etc/jsondb

If no partitions*.json files are found, the database creates a
single partition in the current working directory called "default".

The partitions are reflected in the ephemeral partition as objects
of type "Partition".

A subsequent patch will make it an error to attempt to create an
object of type "Partition". For now, these objects will just be
ignored.

The -base-name and -dbdir arguments to jsondb are now ignored.

Change-Id: Idaed80444e17bd8b9c4f359c47233169263cedd9
Reviewed-by: Jamey Hicks <jamey.hicks@nokia.com>