Split partition handling out into a separate module
This is a brute force attempt to create a jsondbpartition module
which can be used by the daemon, by autotests and by the client for
private partitions.

Most of the classes were largely unmodified by this change (with
the exception of namespace and symbol exports). JsonDbManagedBtree
was somewhat reworked in order not to expose QBtree as part of this
module.

Overall, the number of headers exposed by this module is far too
large currently. Subsequent patches are needed to refine the APIs
of this module, to make all the headers private and to overall
reduce the number of headers provided.

Change-Id: I6deaf32960102f9d3efcb290edd82381597657aa
Reviewed-by: Jamey Hicks <jamey.hicks@nokia.com>