Only sync state numbers of indexes on close.
Currently, whenever the IndexSyncTimer fires (by default 12 seconds
after the last commit), all the indexes of all the object tables are sync'd.

If an index's stateNumber is not up-to-date, the state number is
updated. This behavior causes all index files to have at least 2 pages
written on sync().

This change adds a new JsonDbObjectTable::SyncFlag:
JsonDbObjectTable::SyncStateNumbers. The state number will only be
updated in index files when the partition is closed. This is done so
that when jsondb is restarted it does not reindex the index.

Updated so all auto tests pass.

Fixes Bug #9893.

Change-Id: I677861a6559a8ac3d033d45532ebc2fc5387e404
Reviewed-by: Kevin Simons <kevin.simons@nokia.com>