Fix a case where map/reduce definition updates were not processed
If you created a View, removed the View before removing the source
objects and then recreated the same View again, there would be
old target objects hanging around.

This is an overly complex solution to a simple problem however. We
need to in the future just drop the b-tree files for the views when
they are removed.

Change-Id: I067b76f5e279c5bfce884fcf0f521c7ecd30dc60
Reviewed-by: Jamey Hicks <jamey.hicks@nokia.com>