Fixed a corner case in eager view updates
DBServer::objectsUpdated was being called with an empty change list
because one of the views being updated caused no changes to view
objects. This was not detected as an update of view objects, and so
the wrong code path was chosen.

As a result of this fix, no stateChange notification will be sent if
no objects are updated.

Fixes Bug #11592

Change-Id: I2f0bf4f2711b5ec1e0bc7d88553887bbf8b7ce85
Reviewed-by: Kevin Simons <kevin.simons@nokia.com>