Add new private QJsonDbQueryModel class to client.
This change separates the cachinglistmodel logic from
the implementation of qml-specific aspects of the
jsondbcachinglistmodel.

The code is migrated into new private classes in client and
reused by the imports/jsondb models.
The new classes include functionality of the following
components previously embedded in the qml jsondb plugin:
- JsonDbCachingListModel
- modelutils
- modelcache

In the new setup, the QJsonDbQueryModel class
implements the functionality previously provided by the
src/imports/jsondb JsonDbCachingListModel and
JsonDbCachingListModelPrivate classes.

Modelutils and modelcache helpers are now also part of the
private API in client.
No functional / performance regression is observed
after running models tests and benchmarks on top of this.

The idea is to make QJsonDbQueryModel part of the public
C++ api after a few rounds of reviewes & iterations.

Change-Id: Ia6e2368a4cb7c7485087e714642e00ff686945b7
Reviewed-by: Tapani Mikola <tapani.mikola@nokia.com>