Added commit chains to hbtree
Allow multiple versions of a page to exist so that in memory
rollbacks can be performed to at lease one previous commit. This
is needed in the case that a commit fails half way through because
of out of disk space (for e.g.). Without this the btree
overwrites valid pages.

Change-Id: I7446345c7c373e5ffa0871238a0e61aa7f39fdf0
Reviewed-by: Carlos Duclos <carlos.duclos@nokia.com>
Reviewed-by: Jamey Hicks <jamey.hicks@nokia.com>