Fix for out of space related bugs.
Each partition keeps track of the disk usage, and whenever a commit fails
on that partition, we call that class and check if the failure was
due to OutOfSpace conditions. If so, we mark the partition as
"read-only", which means create, update and remove queries are forbidden
but we will reply to queries.

Change-Id: Ic9eab20eaec37be8875d3dc214d7c71b20b87ba7
Reviewed-by: Jamey Hicks <jamey.hicks@nokia.com>