Discard old versions from changesSince cache.
Each time JsonDbObjectTable::changesSince is called, it reads in the
necessary changes to a cache of the change log.

This changes it to discard any versions from the cache that are more
than JsonDbSettings::changeLogCacheVersions() state changes older than
the one requested in this call to changesSince().

The setting JsonDbSettings::changeLogCacheVersions() defaults to 10.

A separate change should be made to flush the cache of the change log
if it hasn't been used recently.

Fixes Bug #10869

Change-Id: I037416ac032485decc78dcb394bb5ed5885dc1ac
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>
Reviewed-by: Ali Akhtarzada <ali.akhtarzada@nokia.com>