Close indexes after they are added and after syncing.
This changes jsondb so that it only opens indexes when they are
used. When a new index is added, it is opened to examine the state
number and to see if it needs to be reindexed, and then it is closed
again.

In addition, closed indexes remain closed after object tables and
indexes are synched.

Adds a manual test to verify, but the test is too long to run as an
auto test.

Fixes Bug #10327

Change-Id: I518c0b65d6d9c5aad74da95f5ec669c1d5677d84
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>