Prevent talking to oneself
Talking to oneself currently causes the history command to
behave incorrectly, essentially because the following query
expression is not restrictive enough in this case:

... [?to=%self | to=%friend][?from=%self | from=%friend] ...

One way of fixing this would be to have the history command
issue a separate query for each "direction", and then sort the
final result chronologically in the client before reporting.

For now, simply prevent talking to oneself altogether.

Change-Id: I1d554293b878aa6b164002c08fb8d8994f227322
Reviewed-by: Kevin Simons <kevin.simons@nokia.com>