Moved jsondb query parser to a separate class.
Also refactored it slightly. JsonDbQueryTerm::value() no longer contains the
actual value when it is bound to a variable (i.e. [?foo=%bar] - the
JsonDbQueryTerm::value() returns Undefined value, but
JsonDbQueryTerm::variable() returns a name "bar"). It was a nice optimisation
to store variable values inline, but the behaviour was a bit "shaky" for cases
when the variable doesn't resolve immediately (which we have in access control
codepath - [?%owner=%typeDomain] - and we parse the query once and afterwards
just change the bindings map.

An optimisation to properly and safely implement storing values of variables
inline can be a next step after the rest of the refactoring.

Change-Id: I2d67c5e77123bb54cc75c65b1a7d2beca265bf7f
Reviewed-by: Tapani Mikola <tapani.mikola@nokia.com>