Fix for requests sent to private partitions.
When the request to a private partitions that is sent is deleted and another
one is sent, it was possible to hit assert that check the request is valid.
With this patch we no longer need to store request pointers, but
QJsonDbConnection acts as a middle man for all requests to private partitions
and validates requestId to ensure the request object is still alive and is the
correct one.

The patch made with Prasant Ullattil.

Change-Id: Iabbf7ba4dec0456b059dcb2dd3c0f1e3362f4cb7
Reviewed-by: Tapani Mikola <tapani.mikola@nokia.com>