Changed HBtree to take advisory lock on each open of the db file.
This changes btree to hold the advisory lock the whole time the btree file is
kept open and accessed. The current implementation does not support concurrent
access from multiple processes (or even from multiple file descriptors in the
same process) - neither for writes nor for reads. Holding the lock on the file
should ensure that if somehow multiple processes try to open the same database
file, we don't just corrupt the db file, but gracefully fail to open it.

Change-Id: Ib0c6e3aec0e087fedbfbb859758773182e74e46f
Reviewed-by: Ali Akhtarzada <ali.akhtarzada@nokia.com>