Update all the eager views that depend on a source object change
This previous change to this part of the code was to prevent a given
object change from being processed multiple times. However, I put the
break in the wrong level of the loop, so only the first eager view
depending on that source type would be updated, and others would just
be marked as updated without being processed.

Fixes Bug #16440.

Change-Id: I1634a249e4ec6923100950fc5f45e449588b01d2
Reviewed-by: Jing Bai <jing.t.bai@nokia.com>