Ensure default partition on non-removable medium
This patch modifies the behavior introduced by
53df42aee7db1b4f8446c63939d2cb3b7c553078 as follows:

Upon loading a partition definition that has both the
removable and default properties set to true, the daemon
now sets the default property to false instead of exiting.

If all partitions are marked as removable, the daemon creates
a new one (in CWD, assumed to be on a non-removable medium!)
to serve as the default partition.

Change-Id: Ia2911d5847040bef7a638d64e8102eba4c8c60a8
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>