Speed up [?_uuid="undefined"] queries
This query is causing jsondb to read the whole object table even
though it will never match.

If the min or max key value provided to JsonDbUuidQuery::seekToStart
is not a valid UUID, then return false. This will cause the query to
return 0 objects without reading any of the object table.

Change-Id: Ie337cd66cb6d789066cd07847cc8c00ed86d5dce
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>