Fixed subscription to a notification without explicitly specifying _type
If the don't know the _type we cannot figure out the object table the object is
stored in, so for now lets assume it is the main object table. This will still
be broken if the assumption is wrong and the query like [?_uuid="{123}"] is
actually against a view object, but this should be fixed in one of the next
patches where we remove the concept of object tables and store all objects in
one main object table.

Change-Id: Icc259cc857b691acca3f316d7dea29244616ffe3
Reviewed-by: Jamey Hicks <jamey.hicks@nokia.com>