Prevent compilation failure when cross compiling
qmlplugindump is built for the target, and hence should not be executed on
the host. To further complicate matters, it has a hard dependency on
Qt Declarative internals and hence can't be adjusted to being a host tool
since it would effectively require a complete Qt build on the host machine.

Change-Id: I979fdd65039664d267f92b79a6791c220fb73b90
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>