Update the git-archive export options
The .gitattributes and .gitignore files do not need to be present in
packaged sources, as they are for people using the Git repository.
People who download tarballs usually don't reimport into Git -- they
should just download the actual repositories.

Meanwhile, the .tag file is quite useful. Instead of making Git extract
to the useless SHA-1 of the $n string, ask git-archive to store
the SHA-1 of the commit being packaged.

Change-Id: Ibe813cb99490f7b8a267cd2c0c6eca57fea8b0bd
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>